requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.9
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==11.0.3
